cloud-functions-python-shared-code/consumer-b/src/requirements.txt
functions-framework
google-cloud-logging
./dist/shared_code_lib-0.0.0-py3-none-any.whl
#-e ../../shared-code